pyproject.toml
[tool.poetry]
name = "research-agent"
version = "1.0.0"
description = "Its a research AI agent that can do several tasks!"
authors = ["HarshitDawar55 <harshit.dawar55@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.55.2"
langchain-openai = "^0.2.10"
langchain-community = "^0.3.8"
fastapi = "^0.115.5"
isort = "^5.13.2"
requests = "^2.32.3"
logging = "^0.4.9.6"
uvicorn = "^0.32.1"
pydantic = "^2.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
